Move the shader validation logic into a module
